### 2025-09-14 #8980 @editorialbot
[PRE REVIEW]: CosinorAge: Unified Python and Web Platform for Biological Age Estimation from Wearable- and Smartwatch-Based Activity Rhythms
**Submitting author:** @jlohunecke (<a href="http://orcid.org/0000-0003-2579-7637">Jacob Leo Oskar Hunecke</a>)
**Repository:** https://github.com/ADAMMA-CDHI-ETH-Zurich/CosinorAge
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.5
