### 2025-09-14 96d65da
Update ORCID IDs for authors in paper.md
