### 2025-09-14 28894c1
Add DOI entries to bibliography
Added DOI information for various articles in the bibliography.
