+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={11},
   pages={1--24},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1186/1742-4682-11-16}
 }
 
 @article{iv-is,
@@ -16,7 +17,8 @@
   number={6},
   pages={563--572},
   year={1990},
-  publisher={Elsevier}
+  publisher={Elsevier}, 
+  doi = {10.1016/0006-3223(90)90523-5}
 }
 
 @article{danilevicz2024measures,
@@ -27,7 +29,8 @@
   number={1},
   pages={132},
   year={2024},
-  publisher={Springer}
+  publisher={Springer}, 
+  doi = {10.1186/s12874-024-02255-w}
 }
 
 @article{ferreira2019circadian,
@@ -38,7 +41,8 @@
   number={9},
   pages={1184--1189},
   year={2019},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis}, 
+  doi = {10.1080/07420528.2019.1636814}
 }
 
 @article{rock2014daily,
@@ -49,7 +53,8 @@
   number={2},
   pages={290--296},
   year={2014},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi = {10.3109/07420528.2013.843542}
 }
 
 @article{shim2024circadian,
@@ -60,7 +65,8 @@
   number={1},
   pages={146},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London}, 
+  doi = {10.1038/s41746-024-01111-x}
 }
 
 @article{cole1992automatic,
@@ -71,7 +77,8 @@
   number={5},
   pages={461--469},
   year={1992},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/sleep/15.5.461}
 }
 
 @article{fischer2021measuring,
@@ -82,7 +89,8 @@
   number={10},
   pages={zsab103},
   year={2021},
-  publisher={Oxford University Press US}
+  publisher={Oxford University Press US}, 
+  doi = {10.1093/sleep/zsab103}
 }
 
 @article{dupre2023relationship,
@@ -92,7 +100,8 @@
   volume={11},
   pages={1182552},
   year={2023},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA}, 
+  doi = {10.3389/fpubh.2023.1182552}
 }
 
 @article{bakrania2016intensity,
@@ -103,7 +112,8 @@
   number={10},
   pages={e0164045},
   year={2016},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA}, 
+  doi = {10.1371/journal.pone.0164045}
 }
 
 @article{cox1972regression,
@@ -114,7 +124,8 @@
   number={2},
   pages={187--202},
   year={1972},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library}, 
+  doi = {10.1111/j.2517-6161.1972.tb00899.x}
 }
 
 @misc{axivity,
@@ -133,7 +144,8 @@
   number={7},
   pages={738--744},
   year={2014},
-  publisher={American Physiological Society Bethesda, MD}
+  publisher={American Physiological Society Bethesda, MD}, 
+  doi = {10.1152/japplphysiol.00421.2014}
 }
 
 @article{van2013separating,
@@ -144,7 +156,8 @@
   number={4},
   pages={e61691},
   year={2013},
-  publisher={Public Library of Science San Francisco, USA}
+  publisher={Public Library of Science San Francisco, USA}, 
[…]
