### 2025-10-08 #8980 @rkurchin
Authors, please note that JOSS is currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy. For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

In the meantime, something to think about: The paper is currently quite long for JOSS – we generally target 1000 words or less. Often, this means that there is content included in the paper that is more appropriate for the software documentation.
