### 2025-10-11 61fc431
Fix punctuation and spacing in paper summary
Correct punctuation and spacing in the summary section for clarity.
