### 2025-11-24 ae165eb
Add README and docs folder
