+++ README.md
@@ -1,0 +1,295 @@
+# CosinorAge
+
+[![Documentation Status](https://readthedocs.org/projects/cosinorage/badge/?version=latest)](https://cosinorage.readthedocs.io/en/latest/)
+[![PyPI version](https://badge.fury.io/py/cosinorage.svg)](https://badge.fury.io/py/cosinorage)
+[![Python 3.9+](https://img.shields.io/badge/python-3.9+-blue.svg)](https://www.python.org/downloads/)
+[![License: Apache 2.0](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
+[![GitHub Actions](https://github.com/jlohunecke/CosinorAge/workflows/CosinorAge/badge.svg)](https://github.com/jlohunecke/CosinorAge/actions)
+[![DOI](https://zenodo.org/badge/DOI/10.1038/s41746-024-01111-x.svg)](https://doi.org/10.1038/s41746-024-01111-x)
+
+
+An open-source Python package for estimating biological age based on circadian rhythms derived from accelerometer data. The package offers a unified framework that integrates data preprocessing, feature extraction, and predictive modeling of biological age using the CosinorAge biomarker.
+
+## Getting Started
+
+### Prerequisites
+- Python >= 3.9
+- pip (Python package installer)
+- git
+
+### Installation Steps
+
+#### Option 1: Quick Installation via pip (Recommended)
+```bash
+pip install cosinorage
+```
+
+#### Option 2: Manual Installation from Source
+
+1. **Clone the Repository**
+   ```bash
+   git clone https://github.com/yourusername/cosinorage.git
+   cd cosinorage
+   ```
+
+2. **Set Up Virtual Environment**
+   ```bash
+   # Create a new virtual environment
+   python -m venv venv
+
+   # Activate the virtual environment
+   # On Windows:
+   venv\Scripts\activate
+   # On macOS/Linux:
+   source venv/bin/activate
+   ```
+
+3. **Install Dependencies**
+   ```bash
+   # Upgrade pip
+   pip install --upgrade pip
+
+   # Install required packages
+   pip install -r requirements.txt
+   ```
+
+4. **Install the Package**
+   ```bash
+   # Install in development mode
+   pip install -e .
+   ```
+
+### Quick Start
+After installation, you can start using the package:
+
+```python
+import cosinorage
+from cosinorage.datahandlers import GalaxyDataHandler
+
+# Initialize a data handler
+handler = GalaxyDataHandler(galaxy_file_dir='path/to/your/data')
+
+# Compute features
+from cosinorage.features import WearableFeatures, dashboard
+features = WearableFeatures(handler)
+dashboard(features)
+
+# Compute CosinorAge (dummy data)
+from cosinorage.bioages import CosinorAge
+records = [
+    {'handler': handler, 
+     'age': 60, 
+     'gender': 'male', 
+    }
+]
+predictions = CosinorAge(records)
+predictions.get_predictions()
+```
+
+For more detailed examples, please refer to the [examples](examples) directory.
+
+## Package Functionalities
+
+For detailed examples of how to use the package, please refer to the [examples](examples) folder. The package's theoretical background, including references to the underlying research papers, is documented in our [Read the Docs theoretical background section](https://cosinorage.readthedocs.io/en/latest/theoretical_background.html). For comprehensive API documentation, please see the [package reference](https://cosinorage.readthedocs.io/en/latest/cosinorage.html).
+
+### Modular Scheme 
+
+![Package Data Scheme](docs/figs/schema.jpg)
+
+### Data Loading
+
+Depending on the data source, different data handlers are used to perform the necessary preprocessing and data conversion steps. These handlers ensure that the data is consistently formatted as minute-level ENMO data.
+
[…]
