+++ docs/source/cosinorage.bioages.rst
@@ -1,0 +1,17 @@
+cosinorage.bioages Module
+=========================
+
+Module Contents
+---------------
+
+.. automodule:: cosinorage.bioages
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Classes
+-------
+
+.. autoclass:: cosinorage.bioages.cosinorage.CosinorAge
+   :members:
+
