+++ docs/source/index.rst
@@ -1,0 +1,14 @@
+.. CosinorAge documentation master file, created by
+   sphinx-quickstart on Thu Nov  7 17:26:12 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+CosinorAge Documentation
+========================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   cosinorage
+   theoretical_background
