### 2026-01-08 8b41776
Update paper.md
