+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 Every day, millions of people track their steps, sleep, and activity rhythms using smartwatches and fitness trackers. These continuous data streams offer an opportunity to transform routine self-tracking into meaningful health insights that inform biological aging. However, most wearable data tools remain fragmented, proprietary, and inaccessible, limiting translation into actionable knowledge.
 
-`CosinorAge` is an open-source framework that estimates biological age from wearable-derived circadian, physical activity, and sleep metrics. It provides a unified, reproducible Python pipeline for data preprocessing, feature computation, and biological age estimation, with trained model parameters from large-scale datasets such as the UK Biobank. Its companion `CosinorAge Calculator` offers identical functionality via a web interface. Together, they enable transparent, scalable, and personalized health monitoring while bridging digital health and biological aging research..
+`CosinorAge` is an open-source framework that estimates biological age from wearable-derived circadian, physical activity, and sleep metrics. It provides a unified, reproducible Python pipeline for data preprocessing, feature computation, and biological age estimation, with trained model parameters from large-scale datasets such as the UK Biobank. Its companion `CosinorAge Calculator` offers identical functionality via a web interface. Together, they enable transparent, scalable, and personalized health monitoring while bridging digital health and biological aging research.
 
 # Statement of Need
 
@@ -43,9 +43,9 @@ Circadian rhythms play a critical role in maintaining key regulatory systems, in
 
 Wearable devices and smartwatches enable a scalable, non-invasive, and cost-efficient method for digital biomarkers of circadian rhythms, physical activity, and sleep at both individual and population levels. However, most analytic tools focus on isolated metrics or rely on proprietary algorithms, limiting transparency, reproducibility, and their linkage to health outcomes such as biological age. To address this gap, we developed `CosinorAge` [@shim2024circadian], a digital biomarker framework that estimates biological age and healthspan from circadian rest-activity rhythms using wearables.
 
-`CosinorAge` provides a unified, open-source analytic pipeline for processing raw wearable data—from ingestion and preprocessing to feature computation and biological age estimation. While existing tools such as pyActigraphy [@hammad2021pyactigraphy], actipy [@actipy], CosinorPy [@movskon2020cosinorpy], and scikit-digital-health [@adamowicz2022scikit] analyze analyzing specific domains, and GGIR [@ggir] lacks functionalities that link metrics to health outcomes and Python implemention, `CosinorAge` integrates circadian, activity, and sleep domains into a single reproducible framework. Importantly, `CosinorAge` represents a digital *second-order clock* derived from mortality risk rather than chronological age, capturing healthspan-relevant processes with higher precision [@shim2024circadian].
+`CosinorAge` provides a unified, open-source analytic pipeline for processing raw wearable data—from ingestion and preprocessing to feature computation and biological age estimation. While existing tools such as pyActigraphy [@hammad2021pyactigraphy], actipy [@actipy], CosinorPy [@movskon2020cosinorpy], and scikit-digital-health [@adamowicz2022scikit] analyze specific domains, and GGIR [@ggir] lacks functionalities that link metrics to health outcomes and Python implementation, `CosinorAge` integrates circadian, activity, and sleep domains into a single reproducible framework. Importantly, `CosinorAge` represents a digital *second-order clock* derived from mortality risk rather than chronological age, capturing healthspan-relevant processes with higher precision [@shim2024circadian].
 
[…]
