### 2026-01-09 471487e
Update paper.md to ensure adherence to new JOSS requirements
