### 2026-01-09 #8980 @jlohunecke
Thanks for the comments. We have addressed all requests.

1. **Sections added**: We added the "Software Design," "Research Impact Statement," and "AI usage disclosure" sections to the manuscript.
2. **Repository history**: In November 2024, the codebase was first committed from a local development environment to GitHub. As a result, GitHub metrics show that 99.9% of activity is concentrated between November 4-6, 2024, even though the majority of development occurred earlier outside of GitHub. A substantial portion of the line count comes from adding large CSV files for testing purposes (publicly available data).

Please let us know if you have any additional questions.
