### 2026-02-28 #8980 @jlohunecke
Dear JOSS Team,

I hope you’re doing well. I wanted to kindly ask whether you have an approximate timeline for when the first review might start. This would help us plan our resources accordingly.

Best regards,
Jacob

> On 16. Feb 2026, at 12:09, The Open Journals editorial robot ***@***.***> wrote:
>
>
> editorialbot
>  left a comment
> (openjournals/joss-reviews#8980)
>  <https://github.com/openjournals/joss-reviews/issues/8980#issuecomment-3907855585>
> Five most similar historical JOSS papers:
>
> ⚠️ Important note to authors: Please do not directly contact the authors of these papers. If you believe these papers are relevant to your submission, please mention this in your submission or communicate with your handling editor.
>
> Devicely: A Python package for reading, timeshifting and writing sensor data <https://github.com/openjournals/joss-reviews/issues/3679>
> Submitting author: @arianesasso
> Handling editor: @prashjha (Active)
> Reviewers: @luciorq, @djmitche
> Similarity score: 0.6787
>
> SleepECG: a Python package for sleep staging based on heart rate <https://github.com/openjournals/joss-reviews/issues/5411>
> Submitting author: @cbrnr
> Handling editor: @emdupre (Active)
> Reviewers: @tompollard, @richrobe, @peterakirk
> Similarity score: 0.6695
>
> mhealthtools: A Modular R Package for Extracting Features from Mobile and Wearable Sensor Data <https://github.com/openjournals/joss-reviews/issues/2106>
> Submitting author: @philerooski
> Handling editor: @trallard (Retired)
> Reviewers: @Ebedthan, @AKuederle
> Similarity score: 0.6677
>
> BioPsyKit: A Python package for the analysis of biopsychological data <https://github.com/openjournals/joss-reviews/issues/3702>
> Submitting author: @richrobe
> Handling editor: @osorensen (Active)
> Reviewers: @zen-juen, @espenhgn
> Similarity score: 0.6567
>
> activAnalyzer: An R Shiny app to analyse ActiGraph accelerometer data and to implement the use of the PROactive Physical Activity in COPD instruments <https://github.com/openjournals/joss-reviews/issues/4741>
> Submitting author: @pydemull
> Handling editor: @samhforbes (Active)
> Reviewers: @elimillera, @angerhang
> Similarity score: 0.6491
>
> ⚠️ Note to editors: If these papers look like they might be a good match, click through to the review issue for that paper and invite one or more of the authors before considering asking the reviewers of these papers to review again for JOSS.
>
