### 2026-03-16 #8980 @arianesasso
@jromanowska Hey 👋🏼 ! I'd be happy to support with the review. However, I believe this would be my first one for JOSS. Anything else I need to know besides the guidelines??? Thanks!
