### 2026-03-16 #8980 @jromanowska
@arianesasso , thank you for agreeing! The review process will be happening in a separate issue here, on GitHub. It's basically a discussion between you (the reviewer) and the authors, aimed at improving the software. As you can read in the docs, you will get a checklist to guide you through your review. We aim at completing the review within 4 weeks, but this starts when I find at least 2 reviewers. So for now, you can wait or look around in the repo of the submitted software, if you wish.
And do not hesitate to ask me if you have any other questions!
