### 2026-03-23 #8980 @jromanowska
:wave: @angerhang, @limengbinggz, @epiben, @ManuelaS & @elyiorgos, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
