### 2026-03-25 #8980 @angerhang
> 👋 [@angerhang](https://github.com/angerhang), [@limengbinggz](https://github.com/limengbinggz), [@epiben](https://github.com/epiben), [@ManuelaS](https://github.com/ManuelaS) & [@elyiorgos](https://github.com/elyiorgos), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@jromanowska thanks for the invite. Looks like a very interesting submission. Sadly have to say no on this occasion as I am in a middle of job transition.
