### 2026-03-26 #8980 @jromanowska
@angerhang, thank you so much!

:wave: @chanshing, @aidanacquah, would any of you be willing to review this software submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
