### 2026-04-18 #8980 @jlohunecke
Dear JOSS Team,

I hope you are doing well.

I am writing to kindly inquire about the current status of our submission. As we submitted our manuscript on September 1, 2025, we would greatly appreciate any updates regarding its review progress.

Given the extended waiting period, we would also like to ask whether there might be any possibility of advancing the review process.

Thank you very much for your time and consideration.

Best regards,
Jacob

> On 7. Apr 2026, at 09:14, Julia ***@***.***> wrote:
>
>
> jromanowska
>  left a comment
> (openjournals/joss-reviews#8980)
>  <https://github.com/openjournals/joss-reviews/issues/8980#issuecomment-4197191821>
> 👋 @limengbinggz <https://github.com/limengbinggz>, @epiben <https://github.com/epiben>, @ManuelaS <https://github.com/ManuelaS> & @elyiorgos <https://github.com/elyiorgos>, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
