### 2026-05-31 #10457 @pydemull
@jlohunecke, for information I have posted several issues regarding the software documentation: [issue #6](https://github.com/ADAMMA-CDHI-ETH-Zurich/CosinorAge/issues/6); [issue #7](https://github.com/ADAMMA-CDHI-ETH-Zurich/CosinorAge/issues/7); [issue #8](https://github.com/ADAMMA-CDHI-ETH-Zurich/CosinorAge/issues/8). In addition, in the “Wearable Features Computions” section of the tutorial, there is a typo: “To genereate all the visulaizations”. There is another typo in the “Circadian Rhythm Analysis - Non-parametric Analsis” subtitle and the “Sleep Analyis” subtitle.
