### 2026-06-05 d82d596
Fix JOSS reviewer issues for README and paper figures.
Correct NHANES and Galaxy Quick Start examples, repair broken paper figure paths, and avoid special characters that break JOSS rendering.

Co-authored-by: Cursor <cursoragent@cursor.com>
