+++ paper/paper.md
@@ -88,7 +88,7 @@ The Web interface is organized into several sections:
 
 The Calculator tab offers a user-friendly interface, as illustrated in Figure 3. **CosinorAge Calculator** supports BYOD via batch CSV uploads from either single or multiple individuals, with automatic file structure preview for validation (subject to file size limits). Users can configure device type, timestamp format, time zone, and select parameters for analysis. When multi-individual mode is selected, the summary dashboard presents descriptive statistics for all extracted features, a feature correlation matrix, and visual summaries of each metric at the population level.
 
-![Data upload interface & summary dashboard.](figures/calc.png)
+![Data upload interface and summary dashboard.](figures/calc.png)
 
 # Research Impact Statement
 
