### 2026-06-05 848ffb5
Rename paper timeseries figure to avoid GitHub filename issues.
Use figures/timeseries.png instead of timeseries_CA.png so the JOSS paper image path matches the committed asset name.

Co-authored-by: Cursor <cursoragent@cursor.com>
