### 2026-06-05 a171b27
Release v1.0.7 with JOSS reviewer fixes and NHANES pandas compatibility.
Improve installation docs, fix NHANES resampling for object-dtype SAS columns, and bump version for PyPI.

Co-authored-by: Cursor <cursoragent@cursor.com>
