+++ cosinorage/__init__.py
@@ -19,7 +19,7 @@
 # limitations under the License.
 ##########################################################################
 
-__version__ = "1.0.4"
+__version__ = "1.0.7"
 
 from . import bioages, datahandlers, features
 
