### 2026-06-05 2169b6b
Fix Read the Docs build by mocking heavy dependencies.
Use Python 3.11, drop claid/skdh/CosinorPy from doc requirements, and mock them in Sphinx autodoc so RTD does not need gfortran.

Co-authored-by: Cursor <cursoragent@cursor.com>
