### 2026-06-05 b0c3609
Fix typos in README wearable features section.
Correct spelling in visualization instructions and section headings for non-parametric and sleep analysis.

Co-authored-by: Cursor <cursoragent@cursor.com>
