### 2026-06-26 b432b63
Address JOSS NHANES feedback on loading time and CosinorAge demographics.
Optimize minute-level NHANES reads by filtering SEQN before chunk processing, improve progress messaging, document expected PAXMIN scan duration and DEMO-based age/gender lookup, and fix the calculate_measure_time test expectation.

Co-authored-by: Cursor <cursoragent@cursor.com>
