+++ docs/source/requirements.txt
@@ -1,17 +1,13 @@
 sphinx
 furo
 sphinx-autodoc-typehints
-sphinx-rtd-theme
 pandas
-tqdm
 numpy
 matplotlib
 seaborn
 scipy
+tqdm
 scikit-learn
 statsmodels
-scikit-digital-health
-claid
-CosinorPy
 
 git+https://github.com/readthedocs/sphinx-build-compatibility#egg=sphinx-build-compatibility
