+++ .readthedocs.yaml
@@ -9,11 +9,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.9"
-    # You can also specify other tool versions:
-    # nodejs: "19"
-    # rust: "1.64"
-    # golang: "1.19"
+    python: "3.11"
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
@@ -30,3 +26,5 @@ sphinx:
 python:
   install:
     - requirements: docs/source/requirements.txt
+    - method: pip
+      path: .
