### 2026-06-27 e46f94f
Release v1.0.8 with NHANES loading improvements and JOSS documentation updates.
Bump version for PyPI after optimizing minute-level NHANES reads, clarifying load-time expectations, and documenting age/gender lookup for CosinorAge predictions.

Co-authored-by: Cursor <cursoragent@cursor.com>
