+++ setup.py
@@ -6,7 +6,7 @@ long_description = (here / "README.md").read_text(encoding="utf-8")
 
 setup(
     name="cosinorage",
-    version="1.0.7",
+    version="1.0.8",
     description="A package for computing the CosinorAge from raw accelerometer data.",
     long_description=long_description,
     long_description_content_type="text/markdown",
