### 2026-09-06 #10457 @jlohunecke
Hi @arianesasso, just a quick update: all points raised in [issue #13](https://github.com/ADAMMA-CDHI-ETH-Zurich/CosinorAge/issues/13) have now been addressed — paper.md revisions were merged, and the `cosinorage.app` outage mentioned in point 8 is also fixed. Thanks again for the thorough review!

cc: @jromanowska
