### 2026-09-07 f85223b
Fix typos in the documentation
